Weaken "`wordsAndSpaces` works like `words`"
Crucially, `words` and `unwords` do magic in their edge cases which can
lose information. For example `words "fish " == words "fish"`. We do not
want to copy those edge cases, because it will break our way more
important property `unwordsAndSpaces . wordsAndSpaces == id`.

To weaken this property, we need to only consider lists of words whose
length is > 0. To do this in quickcheck, I'm using the type [()] (lists
of unit) as a proxy for a natural number type. Hopefully there's a
better way to do that.